sensors: break off RC handling into its own class
